pyproject.toml
[tool.poetry]
name = "ccrawler"
version = "0.0.1"
description = "A GDPR/Cookie consent crawler"
authors = ["Filip Nyquist <fillerix99@gmail.com>", "Teo Hildebrand <teo...@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
splinter = "^0.14.0"
selenium = "^3.141.0"
loguru = "^0.5.3"
pendulum = "^2.1.2"
langdetect = "^1.0.8"
jsonpickle = "^2.0.0"
deep-translator = "^1.4.2"
Pillow = "^8.2.0"
yandex-translater = "^6.0"
textblob = "^0.15.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
